Bug fixed: Hardcoding sprite = 404 sometimes.
What I was about to do:
"When getting the list, we also get the details of the pokemons and store them to show to the user in case he clicks."

I will harcode the image anyways. It still give errors when doing it in this way because the images are not in the same place everytime and I found no way to do it everytime.